Rename reports event to reports-v1, fix references.
