+++ src/events/reports-v1/generate-data/_build-base-json.js
