+++ src/events/reports-v1/generate-data/_build-timeseries.js
