+++ src/scheduled/gen-reports/index.js
@@ -2,7 +2,7 @@ const arc = require('@architect/functions')
 
 exports.handler = async function generateReports () {
   await arc.events.publish({
-    name: 'reports',
+    name: 'reports-v1',
     payload: {}
   })
 }
