+++ tests/integration/events/reports-v1/expected-results/locations.json
