+++ tests/run-report-tests.js
@@ -1,6 +1,6 @@
 // Run selected report tests.
 
-require('./unit/events/reports/generate-data/_build-timeseries-test.js')
-require('./unit/events/reports/generate-data/_multivalent-record-test.js')
-require('./integration/events/reports/generate-data/_build-base-json-test.js')
-require('./integration/events/reports/index-test.js')
+require('./unit/events/reports-v1/generate-data/_build-timeseries-test.js')
+require('./unit/events/reports-v1/generate-data/_multivalent-record-test.js')
+require('./integration/events/reports-v1/generate-data/_build-base-json-test.js')
+require('./integration/events/reports-v1/index-test.js')
